% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getout}
\alias{getout}
\title{A getout function
This function allows you to extract outlinks for a given user with customized user agent}
\usage{
getout(url = "http://datafireball.com", useragent = "R Statistics")
}
\arguments{
\item{url}{URL that need to analyzed Defaults to http://datafireball.com}
}
\description{
A getout function
This function allows you to extract outlinks for a given user with customized user agent
}
\examples{
getout()
}
\keyword{crawl}
\keyword{html}
\keyword{outlinks}
\keyword{useragent}
\keyword{xml}

